refactor: renamed interfaces/util/index.ts to validation.ts
