+++ src/interfaces/util/validation.ts
